+++ spec/rubocop/cop/lint/ambiguous_regexp_literal_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 RSpec.describe RuboCop::Cop::Lint::AmbiguousRegexpLiteral, :config do
-  context 'with a regexp literal in the first argument' do
+  shared_examples 'with a regexp literal in the first argument' do
     context 'without parentheses' do
       it 'registers an offense and corrects when single argument' do
         expect_offense(<<~RUBY)
@@ -177,5 +177,13 @@ RSpec.describe RuboCop::Cop::Lint::AmbiguousRegexpLiteral, :config do
         end
       end
     end
+  end
+
+  context 'Ruby <= 2.7', :ruby27 do
+    include_examples 'with a regexp literal in the first argument'
+  end
+
+  context 'Ruby >= 3.0', :ruby30 do
+    include_examples 'with a regexp literal in the first argument'
   end
 end