+++ spec/rubocop/cop/style/method_call_with_args_parentheses_spec.rb
@@ -424,6 +424,23 @@ RSpec.describe RuboCop::Cop::Style::MethodCallWithArgsParentheses, :config do
         RUBY
       end
 
+      it 'does not register an offense when hash value omission with parentheses and using modifier form' do
+        expect_no_offenses(<<~RUBY)
+          do_something(value:) if condition
+        RUBY
+      end
+
+      it 'registers and corrects an offense when explicit hash value with parentheses and using modifier form' do
+        expect_offense(<<~RUBY)
+          do_something(value: value) if condition
+                      ^^^^^^^^^^^^^^ Omit parentheses for method calls with arguments.
+        RUBY
+
+        expect_correction(<<~RUBY)
+          do_something value: value if condition
+        RUBY
+      end
+
       it 'does not register an offense when without parentheses call expr follows' do
         expect_no_offenses(<<~RUBY)
           foo value:
@@ -441,6 +458,21 @@ RSpec.describe RuboCop::Cop::Style::MethodCallWithArgsParentheses, :config do
 
         expect_correction(<<~RUBY)
           foo(value:)
+          foo arg
+        RUBY
+      end
+
+      it 'registers an offense using assignment with parentheses call expr follows' do
+        # Require hash value omission be enclosed in parentheses to prevent the following issue:
+        # https://bugs.ruby-lang.org/issues/18396.
+        expect_offense(<<~RUBY)
+          var = foo(value:)
+          foo(arg)
+             ^^^^^ Omit parentheses for method calls with arguments.
+        RUBY
+
+        expect_correction(<<~RUBY)
+          var = foo(value:)
           foo arg
         RUBY
       end